pyproject.toml
[project]
name = "ckanext-scss"
version = "0.0.1"
description = ""
readme = "README.md"
authors = [
    {name = ""}
]
license = {text = "AGPL"}
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

keywords = [ "CKAN", ]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest-ckan",
    "pytest-playwright",
    "pytest-factoryboy",
    "pytest-mock",
]

[project.urls]
Homepage = "https://github.com//ckanext-scss"

[project.entry-points."ckan.plugins"]
scss = "ckanext.scss.plugin:ScssPlugin"


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
filterwarnings = [
        "ignore::sqlalchemy.exc.SADeprecationWarning",
        "ignore::sqlalchemy.exc.SAWarning",
        "ignore::DeprecationWarning",
]
addopts = "--ckan-ini test.ini"
